requirements.txt
pandas>=2.0
python-dotenv>=1.0.0
requests>=2.28.0
cryptography>=41.0.0
matplotlib>=3.5
scipy>=1.10.0
yfinance>=0.2.0
